Rename package of BatchMultiObjectiveTrial.
