+++ optuna/multi_objective/_batch_trial.py
